requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
coreschema==0.0.4
cx_Oracle==8.3.0
Django==5.0.9
django-cors-headers==4.6.0
django-mssql-backend==2.8.1
djangorestframework==3.15.2
drf-yasg==1.21.8
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mssql-django==1.5
mysqlclient==2.2.6
packaging==24.2
psycopg2==2.9.10
pyodbc==5.2.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.2
uritemplate==4.1.1
urllib3==2.2.3
